AuthorizationInterceptor and minor fixes
[ AuthorizationInterceptor and InterceptorConfig ]
Now the check to ensure that the JWT token is present in the HTTP request headers is handled through AuthorizationHandler. This way, it's no longer necessary to add "@RequestHeader("Authorization") String token" to all endpoints parameters

[ Other fixes ]
Now the method that looks for books with the same title is case insensitive. It was possible to add two books with the same title without this fix.